Splits the main app into multiple files
